% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{getplot}
\alias{getplot}
\title{Build a ggplot from the reduced feature matrix}
\usage{
getplot(obj, colouring = c("feature", "clustering", "custom"),
  feature = NULL, k = NULL, colours = NULL, pca.plotarrow = TRUE)
}
\description{
Build a ggplot from the reduced feature matrix
}

