% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{reducedim}
\alias{reducedim}
\title{Reduce the dimensionality of the feature matrix. Multiple methods are available.}
\usage{
reducedim(DT, method = c("PCA", "robPCA", "custom"), variables = NULL,
  retpca = FALSE)
}
\description{
Reduce the dimensionality of the feature matrix. Multiple methods are available.
}

