% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{tabfeatures}
\alias{tabfeatures}
\title{Feature matrix from a list of time series}
\usage{
tabfeatures(tslist, features = c("entropy", "FoAcf", "trend", "seasonality"))
}
\arguments{
\item{tslist}{a list of time series}

\item{features}{a vector of features to compute}
}
\value{
A feature matrix.
}
\description{
\code{tabfeatures} computes the feature matrix from a list of time series
}
\examples{
mylist <- list(sunspot.year)
tabfeatures(mylist)
}

