% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsexplore.R
\docType{package}
\name{tsexplore}
\alias{tsexplore}
\alias{tsexplore-package}
\title{Time Series Feature Extraction}
\description{
The tsfeature package provides methods to extract various features from time series data
}

