% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsfeatures.R
\docType{package}
\name{tsfeatures}
\alias{tsfeatures}
\alias{tsfeatures-package}
\title{Time Series Feature Extraction}
\description{
The tsfeature package provides methods to extract various features from time series data
}

